// Frontend/src/app/Services/Models/family.ts
export class Family {
    village_id: String;
    Approved_family_category: String;
    Identity_number: String;
    Date: Date;
    Name_of_wife: String;
    Name_of_husband: String;
    Address: String;
    Date_of_birth: Date;
    Age_at_the_time_of_marriage: Number;
    Job_status: String;
    Education_level: String;
    Number_of_living_children: String;

    Pregnancy_details__Pregnancy__First: String;
    Pregnancy_details__Pregnancy__Second: String;
    Pregnancy_details__Pregnancy__Third: String;
    Pregnancy_details__Pregnancy__Fourth: String;
    Pregnancy_details__Pregnancy__Fifth: String;

    Pregnancy_details__Result__First: String;
    Pregnancy_details__Result__Second: String;
    Pregnancy_details__Result__Third: String;
    Pregnancy_details__Result__Fourth: String;
    Pregnancy_details__Result__Fifth: String;
  
    Family_planning_methods__First__Method: String;
    Family_planning_methods__First__Date: Date;

    Family_planning_methods__second__Method: String;
    Family_planning_methods__second__Date: Date;
        
    Family_planning_methods__third__Method: String;
    Family_planning_methods__third__Date: Date;

    Family_planning_methods__fourth__Method: String;
    Family_planning_methods__fourth__Date: Date;

    Safeforrubella: String;
    Dateofcervicalmucoustest: String;
    Otherdetails: String;
    numberofyoungchildren: Number;
}